+++ extension/src/ui/options/components/toggle/PreferenceToggle.tsx
@@ -1,0 +1,16 @@
+import { Tpreferences } from "@src/utils/preferences";
+import { usePreferencesContext } from "../../context/preferencesContext";
+import Toggle from "./Toggle";
+
+export default function PreferenceToggle(props: { preference: keyof Tpreferences }) {
+    const preferencesContext = usePreferencesContext();
+
+    return (
+        <Toggle
+            checked={!!preferencesContext.preferences[props.preference]}
+            onChange={(e) => {
+                preferencesContext.setPreference(props.preference, e.currentTarget.checked);
+            }}
+        />
+    );
+}
